Continue if missing packages are in cache

+++ entrypoint.sh
@@ -28,8 +28,12 @@ fi
 # Download rpms
 mkdir -p downloads
 cat rpms.lock | xargs -n256 dnf download \
+  --skip-broken \
   --arch=x86_64 --arch=noarch \
-  --downloaddir=downloads
+  --downloaddir=downloads \
+  2> >(tee download.err >&2)
+
+# Remove old packages
 while read package
 do
   if ! grep "^${package%.*.rpm}$" rpms.lock > /dev/null
@@ -37,6 +41,44 @@ do
     rm -fv downloads/${package}
   fi
 done < <(cd downloads; ls -1 *.rpm)
+
+# Check missing packages
+idDownloaded() {
+  if ls downloads/$1.* > /dev/null 2> /dev/null
+  then
+    return 0
+  fi
+  return 1
+}
+
+while read package
+do
+  if ! idDownloaded ${package}
+  then
+    # Error will be printed
+    continue
+  fi
+  line_num=$(grep -n "${package}" rpms.lock | cut -f1 -d:)
+  echo "::warning file=rpms.lock,line=${line_num},title=Package ${package} missing from the repositories"
+done < <(sed -n 's/^No package \(\S*\) available\.$/\1/p' download.err)
+
+line_num=0
+error=false
+while read package
+do
+  if ! idDownloaded ${package}
+  then
+    echo "::error file=rpms.lock,line=${line_num},title=Package ${package} unavailable"
+    error=true
+  fi
+  (line_num+=1)
+done < rpms.lock
+
+if ${error}
+then
+  echo "Missing packages" >&2
+  exit 1
+fi
 
 rm -rf iso-root/Packages
 mkdir -p iso-root/Packages
